Fixed a few slugging issues; there are still more.
Implemented a rudimentary turn cycling, that needs work. I hate the way this is being done; I need to plan this out more. Also, I need to send out messages to members of the story when an entry is posted.

Implemented some basic controls to limit a user from adding a character to multiple stories. The backend code is a bit gross, but it works.

Also removed two unused actions (whoops).

Closes #86
Closes #83
Closes #88

+++ app/models/characters_story.php
@@ -26,6 +26,13 @@ class CharactersStory extends AppModel {
    * @return boolean True on success
    */
   public function add($story_id, $character_id) {
+    // Characters are only allowed in one story at a time. This needs to be a
+    // validation rule at some point, but that requires a refactor.
+    $count = $this->find('count', array(
+      'conditions' => compact('character_id')
+    ));
+    if ($count !== 0) { return false; }
+
     $this->Behaviors->detach('Deactivatable');
 
     $this->id = $this->field('id', compact('story_id', 'character_id'));
